Refactor: Move SeedNodeReportingService to reporting package
